added see reflections test, no database yet

+++ src/tests/fec_test.h
@@ -8,7 +8,7 @@
 typedef struct{
   int thread_num;
   int crate_num;
-  uint16_t slot_mask;
+  uint32_t slot_mask;
   int update_db;
   int final_test;
   char ft_ids[16][50];
